fix ValueError: high is out of bounds for int32

+++ src/tabpfn_extensions/post_hoc_ensembles/sklearn_interface.py
@@ -84,9 +84,9 @@ class AutoTabPFNClassifier(ClassifierMixin, BaseEstimator):
         rnd = check_random_state(self.random_state)
 
         # Torch reproducibility bomb
-        torch.manual_seed(rnd.randint(0, 2**32 - 1))
-        random.seed(rnd.randint(0, 2**32 - 1))
-        np.random.seed(rnd.randint(0, 2**32 - 1))  # noqa: NPY002
+        torch.manual_seed(rnd.randint(0, 2**32 - 1, dtype=np.int64))
+        random.seed(rnd.randint(0, 2**32 - 1, dtype=np.int64))
+        np.random.seed(rnd.randint(0, 2**32 - 1, dtype=np.int64))  # noqa: NPY002
 
         task_type = (
             TaskType.MULTICLASS if len(unique_labels(y)) > 2 else TaskType.BINARY
@@ -97,8 +97,8 @@ class AutoTabPFNClassifier(ClassifierMixin, BaseEstimator):
             max_time=self.max_time,
             ges_scoring_string=self.ges_scoring_string,
             device=self.device,
-            bm_random_state=rnd.randint(0, 2**32 - 1),
-            ges_random_state=rnd.randint(0, 2**32 - 1),
+            bm_random_state=rnd.randint(0, 2**32 - 1, dtype=np.int64),
+            ges_random_state=rnd.randint(0, 2**32 - 1, dtype=np.int64),
             **self.phe_init_args_,
         )
 
@@ -192,9 +192,9 @@ class AutoTabPFNRegressor(RegressorMixin, BaseEstimator):
         rnd = check_random_state(self.random_state)
 
         # Torch reproducibility bomb
-        torch.manual_seed(rnd.randint(0, 2**32 - 1))
-        random.seed(rnd.randint(0, 2**32 - 1))
-        np.random.seed(rnd.randint(0, 2**32 - 1))  # noqa: NPY002
+        torch.manual_seed(rnd.randint(0, 2**32 - 1, dtype=np.int64))
+        random.seed(rnd.randint(0, 2**32 - 1, dtype=np.int64))
+        np.random.seed(rnd.randint(0, 2**32 - 1, dtype=np.int64))  # noqa: NPY002
 
         self.predictor_ = AutoPostHocEnsemblePredictor(
             preset=self.preset,
@@ -202,8 +202,8 @@ class AutoTabPFNRegressor(RegressorMixin, BaseEstimator):
             max_time=self.max_time,
             ges_scoring_string=self.ges_scoring_string,
             device=self.device,
-            bm_random_state=rnd.randint(0, 2**32 - 1),
-            ges_random_state=rnd.randint(0, 2**32 - 1),
+            bm_random_state=rnd.randint(0, 2**32 - 1, dtype=np.int64),
+            ges_random_state=rnd.randint(0, 2**32 - 1, dtype=np.int64),
             **self.phe_init_args_,
         )
